+++ tradingbot/order_type.py
@@ -2,8 +2,12 @@
 from enum import Enum
 
 
-class Operation(str, Enum):
+class OrderType(str, Enum):
   """Define the different operations in trading."""
 
   BUY = 'buy'
   SELL = 'sell'
+  BUYLIMIT = 'buylimit'
+  SELLLIMIT = 'selllimit'
+  BUYSTOP = 'buystop'
+  SELLSTOP = 'sellstop'